This is a minor reformat and cleanup of #2.
This takes @lgw903's branch and performs a number of minor refactors
and cleanups as per the discussion in #2.

* Reordered Bw and added Eigen alignment.
* Added RNG.hpp to wrap generic RNG implementations.
* Made matrices of TSR publicly accessible.

@mkoval and I had some debate about whether public accessors or
public members were more appropriate here.  In the end, we decided
on public members because without some concept of properties or
similar, it would be extremely difficult to do in-place modifications
of the Bw and transforms via public accessor. We might change this in
the future if we need to do inheritance or enforce well-formed-ness.