Fix locking issues.
This commit moves the error-checking logic out of
TrajectoryExecutor::execute into a new TrajectoryExecutor::validate
method. This allows the locked error-checking logic to be performed when
the lock is not held (outside of TrajectoryExecutor::step), enabling
this to work for both the QueuedTrajectoryExecutor as well as the
unqueued TrajectoryExecutors.

In addition, this introduces a TrajectoryMetadata struct, which should
replace prpy's trajectory Tags.