Refactor JointStateSpace to remove Joint.
This decouples the state space from the underlying DART Joint that was
used to generate the state space. Joint properties (e.g. position
limits) are maintained in a separate JointStateSpace::Properties field,
which is not tied to the original DART Joint.

This also changes the interfaces for the JointStateSpace::getState and
JointStateSpace::setState methods to take in a DART Joint.