Refactor MetaSkeletonStateSpace.
This decouples the state space from the underlying DART MetaSkeleton
that was used to generate the state space. Properties are maintained in
a separate MetaSkeletonStateSpace::Properties field.

This changes the interfaces for the MetaSkeletonStateSpace::getState and
MetaSkeletonStateSpace::setState methods to take in a DART MetaSkeleton.

This also changes the interfaces for many of the constraints and
executors to take in an additional DART MetaSkeleton argument.

The VectorFieldPlanner code is in a separate commit so it can be easily
reverted when the improved VectorFieldPlanner is implemented.