Refactor DART StateSpaces (#278)
* Refactor JointStateSpace to remove Joint.

This decouples the state space from the underlying DART Joint that was
used to generate the state space. Joint properties (e.g. position
limits) are maintained in a separate JointStateSpace::Properties field,
which is not tied to the original DART Joint.

This also changes the interfaces for the JointStateSpace::getState and
JointStateSpace::setState methods to take in a DART Joint.

* Rename MetaSkeletonStateSpaceSaver to MetaSkeletonStateSaver.

Since MetaSkeletonStateSpaces won't have an internal MetaSkeleton,
having a MetaSkeletonStateSpaceSaver doesn't make sense.

* Refactor MetaSkeletonStateSpace.

This decouples the state space from the underlying DART MetaSkeleton
that was used to generate the state space. Properties are maintained in
a separate MetaSkeletonStateSpace::Properties field.

This changes the interfaces for the MetaSkeletonStateSpace::getState and
MetaSkeletonStateSpace::setState methods to take in a DART MetaSkeleton.

This also changes the interfaces for many of the constraints and
executors to take in an additional DART MetaSkeleton argument.

The VectorFieldPlanner code is in a separate commit so it can be easily
reverted when the improved VectorFieldPlanner is implemented.

* Refactor VectorFieldPlanner to work with new MetaSkeletonStateSpaces.

One of the VectorFieldPlanner tests is currently failing, but I figured
it's not worth debugging since the improved VectorFieldPlanner will be
ready soon. This is in a separate commit so it can be easily reverted
when the improved implementation is ready.

* Minor formatting fixes.

* Add hash function for std::pair.

* Address Mike's review comments.

Return const & rather than const.

* Address more of Mike's review comments.

* Revert "Refactor VectorFieldPlanner to work with new MetaSkeletonStateSpaces."

This reverts commit 934ee62d016e527432fd5298f1ee2f16e98b0a93.

* Refactor new VectorFieldPlanner to work with new MetaSkeletonStateSpaces.

* Update CHANGELOG.md

* Pass DART objects as const to state spaces (#297)

* Update changelog; Move #278 to 0.3.0

* Pass DART joint as const variable to state spaces

* Use BallJoint::convertToPositions()

* Address Gilwoo's review comments.

* Update TrajectoryMarker per DART state space refactor

* Fix code style