Speed up tests (#418)
* Speed up test_SE2BoxConstraint in debug mode.

In debug mode, this test now runs with fewer samples and targets. Travis
could previously take 40 seconds to run this test, but should now come
closer to the release mode runtime of 0.4 seconds.

* Speed up parabolic smoother tests.

The default resolution is way too fine, especially since the constraint
we're checking against is always satisfied.